Update RoleAssign.cs
Vanilla Among Us is random. With ToH-E the player roles have a pattern when multiple games are played without breaking the lobby or changing any players. 'MersenneTwister.cs' seems to seed players by date/time but I think that only happens at creation of the lobby, based on my testing. I wanted to try to reseed them between games. This builds for me and in my testing fixed the randomization. I am not a coder, I won't be offended if this isn't used but *please* fix the randomization between games within the same lobby.
Thanks. :)